+++ src/cleanup/mod.rs
@@ -27,6 +27,7 @@ fn do_cleanup(
 
     conn_new_style.update_stations_clickcount()?;
     conn_new_style.remove_unused_ip_infos_from_stationclicks(click_timeout_hours)?;
+    conn_new_style.remove_illegal_icon_links()?;
 
     info!("STATS: {} Checks/Hour, {} Checks/Day, {} Working stations, {} Broken stations, {} to do, deletable {} + {}", checks_hour, checks_day, stations_working, stations_broken, stations_todo, stations_deletable_never_worked, stations_deletable_were_working);
     Ok(())